+++ ChangeLog.md
@@ -3,6 +3,10 @@ Komet Changelog
 This changelog summarizes changes and fixes which are a part of each revision.  For more details on the fixes, refer tracking numbers 
 where provided, and the git commit history.
 
+* 2016/11/07 - 1.35?:
+  * Added mapping qualifier as standard map item field.
+  * Fixed problem with javascript calls taking too much time with lots of auto suggest fields on the page
+
 * 2016/11/?? - 1.34: 
     * Updates to cache clearing code for clearing after a call to a /write method
     * Updates to the VHAT export code to try to get through the reverse proxy